% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add_numbers}
\alias{add_numbers}
\title{Add some numbers}
\usage{
add_numbers(x, y)
}
\arguments{
\item{x}{The first number}

\item{y}{The second number}
}
\value{
The sum
}
\description{
Add some numbers
}
\examples{

add_number(5, 2)

}
